[unique_id] Moved to its own file
